gallery/requirements.txt
Django==1.10.5
django-markdown-deux==1.0.5
django-simple-captcha==0.5.3
markdown2==2.3.3
olefile==0.44
Pillow==4.0.0
pytz==2017.2
six==1.10.0
